# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Elite Dangerous Mining Data Server API
  description: |
    # Elite Dangerous Mining Data Server v2.0
    
    A comprehensive API server for Elite Dangerous mining data with real-time statistics, market analysis, and trading route optimization.
    
    ## Features
    - **Real-time EDDN Data Processing**: Live data stream from Elite Dangerous Data Network
    - **Market Data Analysis**: Comprehensive commodity pricing and trading routes
    - **Mining Statistics**: Detailed mining location analytics and profitability tracking
    - **Performance Monitoring**: Server and API usage statistics
    - **WebSocket Integration**: Real-time updates for dynamic data
    
    ## Data Sources
    - **EDDN (Elite Dangerous Data Network)**: Real-time game data via ZeroMQ
    - **Inara API**: Community-driven market and station data  
    - **EDSM API**: System coordinates and exploration data
    
    ## Rate Limiting
    - Standard endpoints: 100 requests per 15 minutes per IP
    - Statistics endpoints: 50 requests per 15 minutes per IP
    - WebSocket connections: 10 concurrent connections per IP
    
    ## Caching
    - Market data: 5-30 minutes TTL depending on volatility
    - Statistics: 1-10 minutes TTL depending on endpoint
    - System data: 24 hours TTL (rarely changes)
    
    ## Response Formats
    All endpoints support JSON by default. Some endpoints also support CSV export via `?format=csv` parameter.
    
  version: "2.0.0"
  contact:
...
servers:
...
paths:
...
  /api/stats:
    get:
      tags:
        - Statistics
      summary: Global server statistics
      description: |
        Returns comprehensive global statistics including:
        - Total systems, stations, and commodities tracked
        - Mining location statistics
        - EDDN message processing statistics
        - API usage metrics
        - WebSocket connection statistics
        - Server performance indicators
      parameters:
        - name: includeHistory
          in: query
          description: Include historical trend data
          schema:
            type: boolean
            default: false
        - name: timeRange
          in: query
          description: Time range for historical data
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d, 30d]
            default: "24h"
        - name: format
          in: query
          description: Response format
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Global statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalStatistics'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/stats/mining:
    get:
      tags:
        - Statistics
      summary: Mining data statistics and analytics
      description: |
        Returns comprehensive mining statistics including:
        - Mining location counts and types
        - Commodity distribution analysis
        - Profitability metrics
        - Hotspot efficiency data
      parameters:
        - name: commodity
          in: query
          description: Filter by specific commodity
          schema:
            type: string
        - name: system
          in: query
          description: Filter by system name
          schema:
            type: string
        - name: includeEfficiency
          in: query
          description: Include efficiency analysis
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Mining statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MiningStatistics'
...
components:
  schemas:
...
    GlobalStatistics:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            systems:
              type: object
              properties:
                total:
                  type: integer
                withStations:
                  type: integer
                withMiningData:
                  type: integer
                averagePopulation:
                  type: number
            stations:
              type: object
              properties:
                total:
                  type: integer
                withMarketData:
                  type: integer
                withShipyard:
                  type: integer
                withOutfitting:
                  type: integer
            commodities:
              type: object
              properties:
                total:
                  type: integer
                tracked:
                  type: integer
                withPriceData:
                  type: integer
            mining:
              type: object
              properties:
                locations:
                  type: integer
                hotspots:
                  type: integer
                uniqueMaterials:
                  type: integer
            eddn:
              type: object
              properties:
                messagesProcessed:
                  type: integer
                uptime:
                  type: number
                connectionStatus:
                  type: string
            api:
              type: object
              properties:
                totalRequests:
                  type: integer
                averageResponseTime:
                  type: number
                errorRate:
                  type: number
            websocket:
              type: object
              properties:
                activeConnections:
                  type: integer
                totalConnections:
                  type: integer
                messagesDelivered:
                  type: integer
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
...
    MiningStatistics:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            locations:
              type: object
              properties:
                total:
                  type: integer
                byType:
                  type: object
                  properties:
                    asteroidBelts:
                      type: integer
                    planetRings:
                      type: integer
                    hotspots:
                      type: integer
                uniqueSystems:
                  type: integer
            commodities:
              type: object
              properties:
                distribution:
                  type: object
                  additionalProperties:
                    type: number
                    description: Percentage distribution
                topProfitable:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      avgProfitPerHour:
                        type: number
                      locations:
                        type: integer
            profitability:
              type: object
              properties:
                topLocations:
                  type: array
                  items:
                    type: object
                    properties:
                      system:
                        type: string
                      body:
                        type: string
                      material:
                        type: string
                      profitPerHour:
                        type: number
                        description: Credits per hour
                averageProfitPerHour:
                  type: number
                totalMinersActive:
                  type: integer
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
...
    ResponseMetadata:
      type: object
      properties:
        source:
          type: string
          enum: [cache, database, api]
        processingTime:
          type: number
          description: Processing time in milliseconds
        timestamp:
          type: string
          format: date-time
        cacheStatus:
          type: string
          enum: [hit, miss, disabled]
        dataFreshness:
          type: string
          enum: [fresh, moderate, stale]
        version:
          type: string
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          nullable: true
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for enhanced rate limits
security:
  - ApiKeyAuth: []
...